forgot to rename useDetailContext
